feat: Add seeder output to meals index
This commit adds a textarea to the meals index page that displays the meals in the same format as the `MealSeeder`. This makes it easy to copy and paste new meals into the seeder file.

This commit also refactors the `MealSeeder` to use a single variable name for all meal creations.