feat: Refactor training progression to support multiple models
This commit refactors the training progression system to support multiple progression models. It introduces a strategy pattern that allows for different progression models to be used for different exercises.

Previously, the training progression logic was hardcoded in the `TrainingProgressionService`. This made it difficult to add new progression schemes. This refactoring addresses that by introducing a `ProgressionModel` interface and two concrete implementations: `LinearProgression` and `DoubleProgression`.

The `TrainingProgressionService` now determines which progression model to use based on the number of reps in the user's last lift for a given exercise. If the reps are within the 8-12 range, it uses the `DoubleProgression` model. Otherwise, it uses the `LinearProgression` model.

The `LinearProgression` model encapsulates the existing logic of suggesting the next weight based on the user's estimated one-rep max. The `DoubleProgression` model implements a new hypertrophy-focused progression, where the user increases the reps at a given weight until they can complete 12 reps, at which point the weight is increased and the reps are reset to 8.

The following changes were made:
- Created a `ProgressionModel` interface.
- Created `LinearProgression` and `DoubleProgression` classes.
- Refactored `TrainingProgressionService` to use the new progression model system.
- Updated `ProgramController` and `LiftLogController` to use the new `TrainingProgressionService` API.
- Added extensive unit tests for the new progression models.
- Added a feature test to ensure that suggested weights, reps, and sets are correctly passed to the mobile entry view.
- Deleted the old `TrainingProgressionServiceTest` as it is now obsolete.