Revert: Remove iOS zoom prevention JavaScript
Removed the JavaScript snippet previously added to `resources/views/app.blade.php` that attempted to prevent pinch-to-zoom on iOS. This was determined not to be the root cause of the layout overflow issue, which was later addressed by CSS changes to the `.git-log` class. This revert cleans up unnecessary code.